[fix] 322、321、317、314问题修改
