+++ aspnet-core/src/Finance.Application/MakeOffers/AnalyseBoard/Model/SopSecondModel.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 
 namespace Finance.MakeOffers.AnalyseBoard.Model;
